+++ packages/eyes-api/src/Eyes.ts
@@ -1,13 +1,13 @@
+import type * as types from '@applitools/types'
 import * as utils from '@applitools/utils'
-import * as logger from '@applitools/logger'
-import SessionType from './enums/SessionType'
-import StitchMode from './enums/StitchMode'
-import MatchLevel from './enums/MatchLevel'
-import TestResultsStatus from './enums/TestResultsStatus'
-import EyesError from './errors/EyesError'
-import NewTestError from './errors/NewTestError'
-import DiffsFoundError from './errors/DiffsFoundError'
-import TestFailedError from './errors/TestFailedError'
+import {SessionType, SessionTypeEnum} from './enums/SessionType'
+import {StitchModeEnum} from './enums/StitchMode'
+import {MatchLevelEnum} from './enums/MatchLevel'
+import {TestResultsStatusEnum} from './enums/TestResultsStatus'
+import {EyesError} from './errors/EyesError'
+import {NewTestError} from './errors/NewTestError'
+import {DiffsFoundError} from './errors/DiffsFoundError'
+import {TestFailedError} from './errors/TestFailedError'
 import {CheckSettings, CheckSettingsFluent} from './input/CheckSettings'
 import {OCRSettings} from './input/OCRSettings'
 import {VisualLocatorSettings} from './input/VisualLocatorSettings'
@@ -20,55 +20,18 @@ import {OCRRegion} from './input/OCRRegion'
 import {ImageRotation, ImageRotationData} from './input/ImageRotation'
 import {CutProviderData} from './input/CutProvider'
 import {LogHandlerData, FileLogHandlerData, ConsoleLogHandlerData, NullLogHandlerData} from './input/LogHandler'
-import {MatchResult, MatchResultData} from './output/MatchResult'
+import {MatchResultData} from './output/MatchResult'
 import {TestResults, TestResultsData} from './output/TestResults'
-import {TestResultsSummary} from './output/TestResultsSummary'
 import {ValidationInfo} from './output/ValidationInfo'
 import {ValidationResult} from './output/ValidationResult'
 import {SessionEventHandler, SessionEventHandlers, RemoteSessionEventHandler} from './SessionEventHandlers'
-import {RunnerConfiguration, EyesRunner, ClassicRunner} from './Runners'
+import {EyesRunner, ClassicRunner} from './Runners'
 import {Logger} from './Logger'
 
-type EyesCommands<TElement = unknown, TSelector = unknown> = {
-  check(options: {
-    settings?: CheckSettings<TElement, TSelector>
-    config?: Configuration<TElement, TSelector>
-  }): Promise<MatchResult>
-  locate<TLocator extends string>(options: {
-    settings: VisualLocatorSettings<TLocator>
-    config?: Configuration<TElement, TSelector>
-  }): Promise<{[key in TLocator]: Region[]}>
-  extractTextRegions<TPattern extends string>(options: {
-    settings: OCRSettings<TPattern>
-    config?: Configuration<TElement, TSelector>
-  }): Promise<{[key in TPattern]: string[]}>
-  extractText(options: {
-    regions: OCRRegion<TElement, TSelector>[]
-    config?: Configuration<TElement, TSelector>
-  }): Promise<string[]>
-  close(): Promise<TestResults>
-  abort(): Promise<TestResults>
-}
-
-type EyesController<TDriver = unknown, TElement = unknown, TSelector = unknown> = {
-  open: (options: {
-    driver: TDriver
-    config?: Configuration<TElement, TSelector>
-    logger?: logger.Logger
-    on?: (event: string, data?: Record<string, any>) => any
-  }) => Promise<EyesCommands<TElement, TSelector>>
-  getResults: () => Promise<TestResultsSummary>
-}
-
-type EyesSpec<TDriver = unknown, TElement = unknown, TSelector = unknown> = {
+type EyesSpec<TDriver = unknown, TElement = unknown, TSelector = unknown> = types.Core<TDriver, TElement, TSelector> & {
   isDriver(value: any): value is TDriver
   isElement(value: any): value is TElement
   isSelector(value: any): value is TSelector
-  makeEyes(config?: RunnerConfiguration): EyesController<TDriver, TElement, TSelector>
-  getViewportSize(driver: TDriver): Promise<RectangleSize>
-  setViewportSize(driver: TDriver, viewportSize: RectangleSize): Promise<void>
-  closeBatch(options: {batchId: string; serverUrl?: string; apiKey?: string; proxy?: ProxySettings}): Promise<void>
-  deleteTestResults(results: TestResults): Promise<void>
 }
 
 export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
@@ -79,7 +42,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
   private _config: ConfigurationData<TElement, TSelector>
   private _runner: EyesRunner
   private _driver: TDriver
-  private _commands: EyesCommands<TElement, TSelector>
+  private _eyes: types.Eyes<TElement, TSelector>
   private _events: Map<string, Set<(...args: any[]) => any>> = new Map()
   private _handlers: SessionEventHandlers = new SessionEventHandlers()
 
@@ -103,7 +66,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
       this._runner = new ClassicRunner()
       this._config = new ConfigurationData(runnerOrConfig as Configuration<TElement, TSelector>)
     }
-    this._runner.attach(this, config => this._spec.makeEyes(config))
+    this._runner.attach(this, this._spec)
     this._handlers.attach(this)
     this._logger = new Logger({...this._config.logs, label: 'Eyes API'})
   }
@@ -143,7 +106,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
   }
 
   get isOpen(): boolean {
-    return Boolean(this._commands)
+    return Boolean(this._eyes)
   }
   getIsOpen(): boolean {
     return this.isOpen
@@ -226,12 +189,11 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
     }
     if (utils.types.isString(testName)) config.testName = testName
     if (utils.types.has(viewportSize, ['width', 'height'])) config.viewportSize = viewportSize
-    if (utils.types.isEnumValue(sessionType, SessionType)) config.sessionType = sessionType
+    if (utils.types.isEnumValue(sessionType, SessionTypeEnum)) config.sessionType = sessionType
 
-    this._commands = await this._runner.open({
+    this._eyes = await this._runner.makeEyes({
       driver,
       config,
-      logger: this._logger,
       on: (name: string, data?: Record<string, any>) => {
         const globalHandlers = this._events.get('*')
         if (globalHandlers) globalHandlers.forEach(async handler => handler(name, data))
@@ -290,7 +252,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
     if (this._config.isDisabled) return null
     if (!this.isOpen) throw new EyesError('Eyes not open')
 
-    let settings
+    let settings: CheckSettings<TElement, TSelector>
     if (utils.types.isString(checkSettingsOrName)) {
       utils.guard.notNull(checkSettings, {name: 'checkSettings'})
       settings = utils.types.instanceOf(checkSettings, CheckSettingsFluent)
@@ -302,7 +264,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
         : {...checkSettingsOrName}
     }
 
-    const result = await this._commands.check({settings, config: this._config.toJSON()})
+    const result = await this._eyes.check({settings, config: this._config.toJSON()})
 
     return new MatchResultData(result)
   }
@@ -313,7 +275,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
     if (this._config.isDisabled) return null
     if (!this.isOpen) throw new EyesError('Eyes not open')
 
-    return this._commands.locate({settings, config: this._config.toJSON()})
+    return this._eyes.locate({settings, config: this._config.toJSON()})
   }
 
   async extractTextRegions<TPattern extends string>(
@@ -322,26 +284,33 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
     if (this._config.isDisabled) return null
     if (!this.isOpen) throw new EyesError('Eyes not open')
 
-    return this._commands.extractTextRegions({settings, config: this._config.toJSON()})
+    return this._eyes.extractTextRegions({settings, config: this._config.toJSON()})
   }
 
   async extractText(regions: OCRRegion<TElement, TSelector>[]): Promise<string[]> {
     if (this._config.isDisabled) return null
     if (!this.isOpen) throw new EyesError('Eyes not open')
 
-    return this._commands.extractText({regions, config: this._config.toJSON()})
+    return this._eyes.extractText({regions, config: this._config.toJSON()})
   }
 
   async close(throwErr = true): Promise<TestResultsData> {
     if (this._config.isDisabled) return null
     if (!this.isOpen) throw new EyesError('Eyes not open')
-    const results = new TestResultsData(await this._commands.close(), results => this._spec.deleteTestResults(results))
-    this._commands = null
+    const results = new TestResultsData(await this._eyes.close(), options =>
+      this._spec.deleteTest({
+        ...options,
+        serverUrl: this._config.serverUrl,
+        apiKey: this._config.apiKey,
+        proxy: this._config.proxy,
+      }),
+    )
+    this._eyes = null
     if (throwErr) {
-      if (results.status === TestResultsStatus.Unresolved) {
+      if (results.status === TestResultsStatusEnum.Unresolved) {
         if (results.isNew) throw new NewTestError(results)
         else throw new DiffsFoundError(results)
-      } else if (results.status === TestResultsStatus.Failed) {
+      } else if (results.status === TestResultsStatusEnum.Failed) {
         throw new TestFailedError(results)
       }
     }
@@ -354,8 +323,15 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
 
   async abort(): Promise<TestResultsData> {
     if (!this.isOpen || this._config.isDisabled) return null
-    const results = new TestResultsData(await this._commands.abort(), results => this._spec.deleteTestResults(results))
-    this._commands = null
+    const results = new TestResultsData(await this._eyes.abort(), options =>
+      this._spec.deleteTest({
+        ...options,
+        serverUrl: this._config.serverUrl,
+        apiKey: this._config.apiKey,
+        proxy: this._config.proxy,
+      }),
+    )
+    this._eyes = null
     return results
   }
   /** @deprecated */
@@ -573,10 +549,10 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
     this._config.setIsDisabled(isDisabled)
   }
 
-  getMatchLevel(): MatchLevel {
+  getMatchLevel(): MatchLevelEnum {
     return this._config.getMatchLevel()
   }
-  setMatchLevel(matchLevel: MatchLevel) {
+  setMatchLevel(matchLevel: MatchLevelEnum) {
     this._config.setMatchLevel(matchLevel)
   }
 
@@ -666,10 +642,10 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
     this._config.setWaitBeforeScreenshots(waitBeforeScreenshots)
   }
 
-  getStitchMode(): StitchMode {
+  getStitchMode(): StitchModeEnum {
     return this._config.getStitchMode()
   }
-  setStitchMode(stitchMode: StitchMode) {
+  setStitchMode(stitchMode: StitchModeEnum) {
     this._config.setStitchMode(stitchMode)
   }