+++ packages/eyes-api/src/enums/FailureReport.ts
@@ -1,0 +1,6 @@
+export enum FailureReportEnum {
+  IMMEDIATE = 'IMMEDIATE',
+  ON_CLOSE = 'ON_CLOSE',
+}
+
+export type FailureReport = `${FailureReportEnum}`