+++ packages/eyes-api/src/input/AccessibilityMatchSettings.ts
@@ -1,5 +1,5 @@
 import * as utils from '@applitools/utils'
-import AccessibilityRegionType from '../enums/AccessibilityRegionType'
+import {AccessibilityRegionType, AccessibilityRegionTypeEnum} from '../enums/AccessibilityRegionType'
 import {Region, RegionData} from './Region'
 
 export type AccessibilityMatchSettings = {
@@ -11,19 +11,22 @@ export class AccessibilityMatchSettingsData implements Required<AccessibilityMat
   private _settings: AccessibilityMatchSettings = {} as any
 
   constructor(settings: AccessibilityMatchSettings)
+  constructor(region: Region)
   constructor(x: number, y: number, width: number, height: number, type?: AccessibilityRegionType)
   constructor(
-    settingsOrX: AccessibilityMatchSettings | number,
+    settingsOrRegionOrX: AccessibilityMatchSettings | Region | number,
     y?: number,
     width?: number,
     height?: number,
     type?: AccessibilityRegionType,
   ) {
-    if (utils.types.isNumber(settingsOrX)) {
-      return new AccessibilityMatchSettingsData({region: {x: settingsOrX, y, width, height}, type})
+    if (utils.types.isNumber(settingsOrRegionOrX)) {
+      return new AccessibilityMatchSettingsData({region: {x: settingsOrRegionOrX, y, width, height}, type})
+    } else if (!utils.types.has(settingsOrRegionOrX, 'region')) {
+      return new AccessibilityMatchSettingsData({region: settingsOrRegionOrX})
     }
-    this.region = settingsOrX.region
-    this.type = settingsOrX.type
+    this.region = settingsOrRegionOrX.region
+    this.type = settingsOrRegionOrX.type
   }
 
   get region(): Region {
@@ -68,13 +71,13 @@ export class AccessibilityMatchSettingsData implements Required<AccessibilityMat
     return this._settings.type
   }
   set type(type: AccessibilityRegionType) {
-    utils.guard.isEnumValue(type, AccessibilityRegionType, {name: 'type', strict: false})
+    utils.guard.isEnumValue(type, AccessibilityRegionTypeEnum, {name: 'type', strict: false})
     this._settings.type = type
   }
-  getType(): AccessibilityRegionType {
-    return this.type
+  getType(): AccessibilityRegionTypeEnum {
+    return this.type as AccessibilityRegionTypeEnum
   }
-  setType(type: AccessibilityRegionType) {
+  setType(type: AccessibilityRegionTypeEnum) {
     this.type = type
   }