+++ packages/eyes-api/src/input/AccessibilitySettings.ts
@@ -1,5 +1,5 @@
-import AccessibilityLevel from '../enums/AccessibilityLevel'
-import AccessibilityGuidelinesVersion from '../enums/AccessibilityGuidelinesVersion'
+import {AccessibilityLevel} from '../enums/AccessibilityLevel'
+import {AccessibilityGuidelinesVersion} from '../enums/AccessibilityGuidelinesVersion'
 
 export type AccessibilitySettings = {
   level?: AccessibilityLevel