+++ packages/eyes-api/src/input/Configuration.ts
@@ -1,12 +1,12 @@
 import * as utils from '@applitools/utils'
-import SessionType from '../enums/SessionType'
-import StitchMode from '../enums/StitchMode'
-import MatchLevel from '../enums/MatchLevel'
-import BrowserName from '../enums/BrowserName'
-import DeviceName from '../enums/DeviceName'
-import ScreenOrientation from '../enums/ScreenOrientation'
+import {SessionType, SessionTypeEnum} from '../enums/SessionType'
+import {StitchMode, StitchModeEnum} from '../enums/StitchMode'
+import {MatchLevelEnum} from '../enums/MatchLevel'
+import {BrowserTypeEnum} from '../enums/BrowserType'
+import {DeviceNameEnum} from '../enums/DeviceName'
+import {ScreenOrientationEnum} from '../enums/ScreenOrientation'
 import {AccessibilitySettings} from './AccessibilitySettings'
-import {DesktopBrowserInfo, ChromeEmulationInfo, IOSDeviceInfo} from './RenderInfo'
+import {DesktopBrowserInfo, ChromeEmulationInfo, IOSDeviceInfo, ChromeEmulationInfoLegacy} from './RenderInfo'
 import {CutProvider} from './CutProvider'
 import {LogHandler} from './LogHandler'
 import {DebugScreenshotProvider} from './DebugScreenshotProvider'
@@ -17,7 +17,7 @@ import {BatchInfo, BatchInfoData} from './BatchInfo'
 import {PropertyData, PropertyDataData} from './PropertyData'
 import {ImageMatchSettings, ImageMatchSettingsData} from './ImageMatchSettings'
 
-type RenderInfo = DesktopBrowserInfo | ChromeEmulationInfo | IOSDeviceInfo
+type RenderInfo = DesktopBrowserInfo | ChromeEmulationInfo | IOSDeviceInfo | ChromeEmulationInfoLegacy
 
 type ConfigurationSpec<TElement, TSelector> = {
   isElement(element: any): element is TElement
@@ -91,7 +91,7 @@ export type ClassicConfiguration<TElement = unknown, TSelector = unknown> = {
 export type VGConfiguration = {
   /** @undocumented */
   concurrentSessions?: number
-  browsersInfo?: RenderInfo[]
+  browsersInfo?: (DesktopBrowserInfo | ChromeEmulationInfo | IOSDeviceInfo)[]
   visualGridOptions?: Record<string, any>
   layoutBreakpoints?: boolean | number[]
   disableBrowserFetching?: boolean
@@ -104,7 +104,8 @@ export type Configuration<TElement = unknown, TSelector = unknown> = GeneralConf
   VGConfiguration
 
 export class ConfigurationData<TElement = unknown, TSelector = unknown>
-  implements Required<Configuration<TElement, TSelector>> {
+  implements Required<Configuration<TElement, TSelector>>
+{
   protected readonly _spec: ConfigurationSpec<TElement, TSelector>
 
   private _config: Configuration<TElement, TSelector> = {}
@@ -116,83 +117,6 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     for (const [key, value] of Object.entries(config)) {
       ;(this as any)[key] = value
     }
-  }
-
-  /** @internal */
-  get general(): Required<GeneralConfiguration> {
-    return utils.general.toJSON(this, [
-      'logs',
-      'debugScreenshots',
-      'agentId',
-      'apiKey',
-      'serverUrl',
-      'proxy',
-      'isDisabled',
-      'connectionTimeout',
-      'removeSession',
-      'remoteEvents',
-    ])
-  }
-
-  /** @internal */
-  get open(): Required<OpenConfiguration> {
-    return utils.general.toJSON(this, [
-      'appName',
-      'testName',
-      'displayName',
-      'viewportSize',
-      'sessionType',
-      'properties',
-      'batch',
-      'defaultMatchSettings',
-      'hostApp',
-      'hostOS',
-      'hostAppInfo',
-      'hostOSInfo',
-      'deviceInfo',
-      'baselineEnvName',
-      'environmentName',
-      'branchName',
-      'parentBranchName',
-      'baselineBranchName',
-      'compareWithParentBranch',
-      'ignoreBaseline',
-      'saveFailedTests',
-      'saveNewTests',
-      'saveDiffs',
-      'dontCloseBatches',
-    ])
-  }
-
-  /** @internal */
-  get check(): Required<CheckConfiguration> {
-    return utils.general.toJSON(this, ['sendDom', 'matchTimeout', 'forceFullPageScreenshot'])
-  }
-
-  /** @internal */
-  get classic(): Required<ClassicConfiguration<TElement, TSelector>> {
-    return utils.general.toJSON(this, [
-      'waitBeforeScreenshots',
-      'stitchMode',
-      'hideScrollbars',
-      'hideCaret',
-      'stitchOverlap',
-      'scrollRootElement',
-      'cut',
-      'rotation',
-      'scaleRatio',
-    ])
-  }
-
-  /** @internal */
-  get vg(): Required<VGConfiguration> {
-    return utils.general.toJSON(this, [
-      'concurrentSessions',
-      'browsersInfo',
-      'visualGridOptions',
-      'layoutBreakpoints',
-      'disableBrowserFetching',
-    ])
   }
 
   /** @undocumented */
@@ -339,10 +263,10 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
   set sessionType(sessionType: SessionType) {
     this._config.sessionType = sessionType
   }
-  getSessionType(): SessionType {
-    return this.sessionType
+  getSessionType(): SessionTypeEnum {
+    return this.sessionType as SessionTypeEnum
   }
-  setSessionType(sessionType: SessionType): this {
+  setSessionType(sessionType: SessionTypeEnum): this {
     this.sessionType = sessionType
     return this
   }
@@ -779,10 +703,10 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     this.defaultMatchSettings = defaultMatchSettings
     return this
   }
-  getMatchLevel(): MatchLevel {
-    return this.defaultMatchSettings?.matchLevel
+  getMatchLevel(): MatchLevelEnum {
+    return this.defaultMatchSettings?.matchLevel as MatchLevelEnum
   }
-  setMatchLevel(matchLevel: MatchLevel): this {
+  setMatchLevel(matchLevel: MatchLevelEnum): this {
     if (!this.defaultMatchSettings) this.defaultMatchSettings = {}
     this.defaultMatchSettings.matchLevel = matchLevel
     return this
@@ -861,13 +785,13 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     return this._config.stitchMode
   }
   set stitchMode(stitchMode: StitchMode) {
-    utils.guard.isEnumValue(stitchMode, StitchMode, {name: 'stitchMode'})
+    utils.guard.isEnumValue(stitchMode, StitchModeEnum, {name: 'stitchMode'})
     this._config.stitchMode = stitchMode
   }
-  getStitchMode(): StitchMode {
-    return this.stitchMode
+  getStitchMode(): StitchModeEnum {
+    return this.stitchMode as StitchModeEnum
   }
-  setStitchMode(stitchMode: StitchMode): this {
+  setStitchMode(stitchMode: StitchModeEnum): this {
     this.stitchMode = stitchMode
     return this
   }
@@ -995,10 +919,10 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     return this
   }
 
-  get browsersInfo(): RenderInfo[] {
+  get browsersInfo(): (DesktopBrowserInfo | ChromeEmulationInfo | IOSDeviceInfo)[] {
     return this._config.browsersInfo
   }
-  set browsersInfo(browsersInfo: RenderInfo[]) {
+  set browsersInfo(browsersInfo: (DesktopBrowserInfo | ChromeEmulationInfo | IOSDeviceInfo)[]) {
     utils.guard.isArray(browsersInfo, {name: 'browsersInfo'})
     this._config.browsersInfo = browsersInfo
   }
@@ -1006,7 +930,9 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     return this.browsersInfo
   }
   setBrowsersInfo(browsersInfo: RenderInfo[]): this {
-    this.browsersInfo = browsersInfo
+    this.browsersInfo = browsersInfo.map(browserInfo =>
+      utils.types.has(browserInfo, 'deviceName') ? {chromeEmulationInfo: browserInfo} : browserInfo,
+    )
     return this
   }
   addBrowsers(...browsersInfo: RenderInfo[]) {
@@ -1014,18 +940,22 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
       utils.guard.isObject(browserInfo, {name: `addBrowsers( arg${index} )`})
     }
     if (!this.browsersInfo) this.browsersInfo = []
-    this.browsersInfo.push(...browsersInfo)
+    this.browsersInfo.push(
+      ...browsersInfo.map(browserInfo =>
+        utils.types.has(browserInfo, 'deviceName') ? {chromeEmulationInfo: browserInfo} : browserInfo,
+      ),
+    )
     return this
   }
   addBrowser(browserInfo: RenderInfo): this
-  addBrowser(width: number, height: number, name?: BrowserName): this
-  addBrowser(browserInfoOrWidth: RenderInfo | number, height?: number, name: BrowserName = BrowserName.CHROME) {
+  addBrowser(width: number, height: number, name?: BrowserTypeEnum): this
+  addBrowser(browserInfoOrWidth: RenderInfo | number, height?: number, name: BrowserTypeEnum = BrowserTypeEnum.CHROME) {
     if (utils.types.isObject(browserInfoOrWidth)) return this.addBrowsers(browserInfoOrWidth)
     else return this.addBrowsers({width: browserInfoOrWidth, height, name})
   }
-  addDeviceEmulation(deviceName: DeviceName, screenOrientation = ScreenOrientation.PORTRAIT) {
+  addDeviceEmulation(deviceName: DeviceNameEnum, screenOrientation = ScreenOrientationEnum.PORTRAIT) {
     if (!this.browsersInfo) this.browsersInfo = []
-    this.browsersInfo.push({deviceName, screenOrientation})
+    this.browsersInfo.push({chromeEmulationInfo: {deviceName, screenOrientation}})
     return this
   }