+++ packages/eyes-api/src/output/TestResults.ts
@@ -1,13 +1,14 @@
+import type * as types from '@applitools/types'
 import * as utils from '@applitools/utils'
 import type {Mutable} from '@applitools/utils'
-import TestResultsStatus from '../enums/TestResultsStatus'
+import {TestResultsStatus, TestResultsStatusEnum} from '../enums/TestResultsStatus'
 import {RectangleSize, RectangleSizeData} from '../input/RectangleSize'
 import {TestAccessibilityStatus} from './TestAccessibilityStatus'
 import {SessionUrls, SessionUrlsData} from './SessionUrls'
 import {StepInfo, StepInfoData} from './StepInfo'
 
 export type TestResults = {
-  readonly id?: string
+  readonly testId?: string
   readonly name?: string
   readonly secretToken?: string
   readonly status?: TestResultsStatus
@@ -41,24 +42,27 @@ export type TestResults = {
 
 export class TestResultsData implements Required<TestResults> {
   private _results: Mutable<TestResults> = {} as any
-  private readonly _deleteTestResults: (result: TestResults) => Promise<void>
+  private readonly _deleteTest: (options: {testId: string; batchId: string; secretToken: string}) => Promise<void>
 
   /** @internal */
-  constructor(results?: TestResults, deleteTestResults?: (result: TestResults) => Promise<void>) {
-    this._deleteTestResults = deleteTestResults
+  constructor(
+    results?: types.Results.TestResult,
+    deleteTest?: (options: {testId: string; batchId: string; secretToken: string}) => Promise<void>,
+  ) {
+    this._deleteTest = deleteTest
     if (!results) return this
     this._results = results instanceof TestResultsData ? results.toJSON() : results
   }
 
-  get id(): string {
-    return this._results.id
+  get testId(): string {
+    return this._results.testId
   }
   getId(): string {
-    return this.id
+    return this.testId
   }
   /** @deprecated */
   setId(id: string) {
-    this._results.id = id
+    this._results.testId = id
   }
 
   get name(): string {
@@ -86,11 +90,11 @@ export class TestResultsData implements Required<TestResults> {
   get status(): TestResultsStatus {
     return this._results.status
   }
-  getStatus(): TestResultsStatus {
-    return this.status
+  getStatus(): TestResultsStatusEnum {
+    return this.status as TestResultsStatusEnum
   }
   /** @deprecated */
-  setStatus(status: TestResultsStatus) {
+  setStatus(status: TestResultsStatusEnum) {
     this._results.status = status
   }
 
@@ -381,12 +385,12 @@ export class TestResultsData implements Required<TestResults> {
   }
 
   isPassed(): boolean {
-    return this._results.status === TestResultsStatus.Passed
+    return this._results.status === TestResultsStatusEnum.Passed
   }
 
   async delete(): Promise<void> {
-    if (!this._deleteTestResults) return
-    return this._deleteTestResults(this._results)
+    if (!this._deleteTest) return
+    return this._deleteTest({testId: this.testId, batchId: this.batchId, secretToken: this.secretToken})
   }
   /** @deprecated */
   async deleteSession(): Promise<void> {